requirements.txt
fastapi
uvicorn
google-adk==1.14.0
langchain-community
wikipedia